+++ source/ray_tracing/ray.c
@@ -9,12 +9,12 @@ t_ray	ray_(t_point3 origin, t_vec3 direction)
 	return (r);
 }
 
-t_point3	at(const t_ray *r, double t)
+t_point3	at(const t_ray *r, float t)
 {
 	return (add(r->origin, multiply(r->direction, t)));
 }
 
-t_vec3	new_ray_dir(t_camera *cam, double u, double v)
+t_vec3	new_ray_dir(t_camera *cam, float u, float v)
 {
 	t_vec3	ret;
